1，机器码授权

+++ wx/QGuiCharts.cpp
@@ -226,7 +226,7 @@ void QGuiCharts::UpdateCharts()
 		{
 			//yabao_shuyinglimit_item->setText(QString::number((int)(data->totols_*(1.0 - data->profit_*0.01) - (yabao_shuyingovertop)) / (int)gfit));
 			double A = std::abs(yabao_shuyingovertop / gfit);
-			yabao_shuyinglimit_item->setText(QString::number(A+A/gfit));
+			yabao_shuyinglimit_item->setText(QString::number((int)(A+A/gfit)));
 		}
 		else
 		{
